Sort of major changes to web-server.py
It is now 127.0.0.1:4200 instead of localhost:8080, I'm not sure why I did that. If you want to host it on your IP, on Mac, go into System Preferences, Network, and you should see your IP address when connected to a network. Copy it, then go onto lines: 66, 76, and 99 and change it to your IP.